requirements.txt
azure-storage-blob
azure-functions

functions-framework==3.*
# google-cloud-storage
# google-cloud-logging
flask-expects-json
python-dotenv
ciso8601